requirements.txt
pydantic<2
fastapi
aiohttp
oauthlib
python-dotenv
uvicorn